+++ src/types/ProductSearchQuery.ts
@@ -35,6 +35,10 @@ export function queryParamsFromProductSearchQuery(
             attribute.values.join(`,`);
           break;
 
+        case 'attributes:not':
+          acc[`filters:not[${attribute.key}]`] = attribute.values.join(`,`);
+          break;
+
         case 'boolean':
           acc[`filters[${attribute.key}]`] = attribute.value ? 'true' : 'false';
           break;